Apply single responsibility principle to frontend component (#99)
Client class performs the parsing of inputs, and functionalities that
should have been part of the main program.

This violates the single responsibility principle as every module or
class should only be responsible for a single part of the
functionality. By delegating the parsing of strings to another module,
this allows easier addition and modification of new parsers and
functionalities.

Let's add a new package parser to improve the cohesiveness and
maintainability of the codebase.