TestUtil: add file content comparison method (#122)
Unlike diff tools, Assert.assertEquals does not provide useful 
information when the content in 2 File objects are not the same. 

This makes troubleshooting difficult. As without a clear indication, 
it is hard to trace the errors.

Let's implement the functionality of printing line differences by 
adding the method TestUtil#compareFileContents.