Increase parsing capabilities using third-party library (#128)
The program expects arguments to be given in a pair, in the form
"-argument value".

This design is not flexible as it does not allow inputs with no value
or multiple values. For example, -incremental and -config file1 file2
file3.

Let's use third-party argument parser library, ArgParse4J to increase
flexibility and functionality. It provides many quality-of-life
features, such as, flexible arguments as well as datatype conversion.