Simplify dashboard template for summary view (#129)
There are 2 main templates, which are loaded separately to process the
information when it is required. This requires the same templates to be
generated multiple times for the data, and greatly increase the size of the
generated report unnecessarily. Ideally, this should only contains data for 
the report.

Let's split up the report generation process and dashboard.

This PR supports the display of the commit history. Further work on the panel
for authorship information will be addressed in another PR.