Dashboard revamp (#163)
There are 2 main templates loaded separately to process the information
from the generated report. This requires the same template files to be
regenerated multiple times, which greatly increases the size of the
generated report unnecessarily. Ideally, it should only contains the
data for the report.

Also, as the dashboard uses vue for its templating engine mostly, let's
better utilize the framework to restructure the dashboard into a single
page application which loads content from the report dynamically.