[#182, #183] FileInfoExtractor: decode non-UTF-8 characters (#180)
FileInfoExtractor#generateFileInfo(String, String)
and FileInfoAnalyzer#isReused(String,String) use
Files#newBufferedReader(Path) to construct the BufferedReader for
reading of files.

This constructed BufferedReader uses the default UTF-8 Charset[1]
and a default CharsetDecoder[2] to decode the bytes
read from files into Strings.[3]

As some files being analyzed may contain special characters that are
not from the UTF-8 character set[3], it causes the BufferedReader to
throw a java.nio.charset.MalformedInputException.

As a result, it creates an incomplete FileInfo in
FileInfoExtractor#generateFileInfo(String, String) process, which causes the
program to crash when analysing the file in 
FileInfoAnalyzer#aggregateBlameAuthorInfo(RepoConfiguration, FileInfo).
This is due to the number of lines analyzed in git blame being greater than
the number of lines inside FileInfo. It will also cause the analysis of
some white-listed files to be skipped inside
FileInfoAnalyzer#isReused(String,String).

Let's update FileInfoExtractor#generateFileInfo(String, String)
and FileInfoAnalyzer#isReused(String,String) to construct the
BufferedReader with an java.io.FileReader instead.

The java.io.FileReader extends java.io.InputStreamReader,
which use the same default UTF-8 Charset. However, it will call 
StreamDecoder#forInputStreamReader(InputStream, Object, CharsetDecoder)
, which will set the CharsetDecoder to replace the malformed input
instead of throwing an error[5], thus allowing the FileInfo to be
constructed with the correct number of lines.

[1] Javadocs of Charset:
https://docs.oracle.com/javase/8/docs/api/java/nio/charset/Charset.html

[2] Javadocs of CharsetDecoder:
https://docs.oracle.com/javase/8/docs/api/java/nio/charset/CharsetDecoder.html

[3] Source code of Files.newBufferedReader(Path, Charset):
http://hg.openjdk.java.net/jdk8u/jdk8u/jdk/file/be30cb2a3088/src/share/classes/java/nio/file/Files.java#l2780

[4] Example of a file with non-UTF-8 characters inside:
https://github.com/CS2103JAN2018-F11-B4/main/blob/master/collated/functional/ZhangYijiong.md

[5] Source code of StreamDecoder setting malformed input to be replaced:
http://hg.openjdk.java.net/jdk8/jdk8/jdk/file/687fd7c7986d/src/share/classes/sun/nio/cs/StreamDecoder.java#l230