Fix dashboard padding function (#187)
When displaying contributors with no commit history, the dashboard will
crash and stop rendering ramp charts for subsequent repositories.

This is caused by the padding of the ramp charts. The padding is added
to the ramp chart to align the charts across different users and
projects.

The amount of padding required is calculated from the `sinceDate` of
the first and last date entry of the user. This becomes problematic
when the user have no date entries at all, causing the system to run
into problem, and hence, skipping the ramp chart generation for the
particular project.

To fix this, let's alter the padding function to take into the edge
case of a empty date range, and artifically fill up the whole timeline
with generated empty dates.