[#194] Fix invalid path when accessing jar resources (#196)
RepoSense v1.1-rc2 executed as a JAR throws an InvalidPathException
when accessing templateZip.zip.

This is because the location of the templateZip.zip cannot be located
by its path on the disk but by its relative path in the JAR file. 

So let's change the way to retrieve templateZip.zip from 
RepoSense.class.getClassLoader().getResource("templateZip.zip") to 
RepoSense.class.getResourceAsStream("/templateZip.zip").