[#206] Fix different branches of same repository using same display name (#207)
In the .csv config file, we may supply the same organization
and repository, but different branches for different authors.

In CsvParser, these will be separated into two different
RepoConfigurations, and analysis will be run separately, generating
separate reports.

However, as we set the display name for each RepoConfiguration to
only be the 'organization' and 'repository', these different
RepoConfigurations will override each other's report as they are
being generated into the same folder.

Let's update RepoConfiguration's display name to include the
branch as well. This will cause the report to be generated into a
different folder for different branches.