[#203] Authorship: incorrect line authorship when analyzing date range (#205)
When a date range is supplied, the FileInfoAnalyzer will only
git blame on the range of commits that fall within the given
date range.

This may cause the authorship of certain lines that have not
been touched at all during the date range to be incorrectly
authored to a random author that commited in that date range.

Let's update the git blame command to run the git blame from
initial to the latest commit of that date range. In addition, let's
also update FileInfoExtractor to only track lines that were added
within the date range, and add them into the final authorship count.