Fix empty ramps when overriding RepoConfiguration (#243)
When a repo has a _reposense/config.json configuration file,
the RepoConfiguration will be reset to the new config.json file,
and all authors are reset as well.

However, these reset authors does not have their author aliases
or ignore glob list set inside them. This may cause some of the
commit info to be missing as we are filtering commits by author
names and aliases.

Let's fix this by updating the
RepoConfiguration#update(StandAloneConfig) to set the
author aliases and ignore glob list for each author inside the
config.json file.