Dashboard: determine date range only after all repo data are loaded (#227)
The dashboard as of v1.1.1 is loaded dynamically after each
`commits.json` is loaded.

If the filter date range is not defined, the range will be determined
by the first and last commits of that particular repository. This start
and end dates will then be set as the filter dates to be used for later
rendering of the other repositories.

However, this leads to a problem when the repositories are of a
different date range, if the earlier repository is loaded first, the
date range will not be able to display the ramps in the later
repository, vice versa.

To fix this issue, let's only determine the date ranges after all the
`commits.json` of the repositories is loaded.