[#223] Dashboard: sort by authors by name within repo  (#233)
The current dashboard as of v1.1.2 sorts the users in the repository
by the number of contributions they have made for the project in that
time period.

This however puts an unnecessary focus on the contribution LoC which
may not be a true reflection of the users' contribution. 

Let's change the default view to sort the users by their names instead.