[#224] Dashboard: restore view state using query parameters (#160)
The summary dashboard plays an important role in highlighting and
filtering the report data to provide the most relevant view for the
users.

Permalinks was a feature implemented in the previous dashboard that
enables the users to quickly save the state of the dashboard to quickly
return to it when needed.

While useful, this feature was not core to the usage of the dashboard,
hence left out during the feature for feature reimplementation during
the dashboard structural changes.

Lets implement this permalink feature in the main vue module. New
fields will be emitted from their individual modules into the global
object.

The summary module will read the information from the global object
then load the filtered view accordingly.

Implementing it in the main modules also provides us the option of
opening up other modules for the permalink features if needed in the
future.