[#209] Dashboard: hide overflow elements (#251)
"label" tag seems to be greater than the height allocated to it by the
render engine. This results in the "scrollbar" arrows popping up to
show the users the "full content".

This can be distracting for users when they are trying to navigate the
site. Let's remove this scrollbar arrows by applying the "overflow-y:
hidden" CSS setting to the filter labels.