[#189] Dashboard: fix ramps z-index value (#255)
Ramps representing very huge commits will be rendered as a very
humongous triangle. This display blocks up the commits in the earlier
periods, hence causing them to be unclickable.

To fix this, let's render the ramps such that the earlier ramps will
have a larger z-index value to place it on top of the more recent
ramps. This ensures that users will be able to interact with all the
ramps.