Add tests to RepoConfiguration#merge (#279)
AuthorConfigCsvParser#getRepoConfiguration should return existing
RepoConfiguration from a list of RepoConfigurations. Otherwise it adds
a newly created RepoConfiguration into the list and returns it.

However, due to a bug, any RepoConfigurations existing or not are added
to the list of RepoConfigurations causing duplicates.

Let's add test to RepoConfiguration#merge to prevent bugs.