[#277] Make authors configuration completely optional (#294)
Author's information is compulsory in configuration; RepoSense requires
the user to configure an author-config.csv or have the repository owner
put up their own configuration.

This, however, creates a hurdle for the new users to overcome as they
would need to build up the configuration before trying out our tool.

Let's improve the first impression of our tool by making authors
configuration completely optional. All the authors found in history
will be added to the analysis in the event that no author configuration
was found for the repository.