[#289] Code view: tweak code packing algorithm (#292)
Untouched code segments that are smaller than minimum lines are
displayed together with the touched code. This distorts the truth as
these code do not belong to the author and may lead to unnecessary
disputes.

Let's encapsulate untouched code in the expand box only if >5 lines but
still differentiate them using color irregardless.