[#301] Hotfix: report folder not generated when failed to analyze repo (#302)
When we fail to clone/checkout a config's repository, we will
skip analyzing that repository, thus not generating its report folder.

However, as the config is still inside the config list, we will still add
that 'failed' config into summary.json. This will cause the dashboard
to crash as it tries to load a report folder which does not exist.

Let's update ReportGenerator to generate the report folder with
'empty' reports when failing to clone/checkout the repository, so as
to allow the dashboard to be viewed, and also allow the user to identify
errors in which repository config info given.