[#85] Rename functional to systemtest (#306)
Functional folder is a skeleton of an end-to-end testing infrastructure.

However the name is not intuitive enough to imply it so. Also, the functional
test is placed in a separate package of the 'src' module instead of the 'test' 
module, making IntelliJ unable to find the classes and resources to run it.

Let's rename functional to system test.