[#325] Replace the usage of findstr and grep (#324)
We use terminal functions, findstr and grep, to filter the results of
command execution.

This requires us to append os dependent command and also subject to the
limitation of these libraries.

For example, findstr does not support proper regex and has no limiting
quantifier support. This can exposes us to problems when searching for 
patterns that cannot be resolved with the regex of findstr.

Let's replace the use of this 2 functions with our own regex-based text
filtration method, StringsUtil#filterText.