[#342] Fix inconsistent length of contribution bar (#346)
Length of the contribution bars varies, dependent on the set of
displayed authors.

This should not be the case. The scale used for the bar should be fixed
for a given period irrespective of which authors are displayed. Search
filters should simply filter out certain authors without changing
anything else.

Let's fixed this by performing the calculations on all authors'
calculation instead of filtered.
Also, since this value should not change in subsequent events, let's
only perform this calculation once and store the value in a global
constant.