[#366] Fix NullPointerException when accessing unfilled displayName in StandaloneConfig (#368)
In our user guide, we allow most of the fields under the authors'
configuration to be optional other than the GitHub Id. Particularly
`Display Name`, which is suppose to use GitHub Id value when it is not
provided. However, when the display name is not filled in the
standalone config, the program crashes due to a NullPointerException.

This is because standalone config adopts the json format. When a field
is not filled in, in the json file, it will be initialized as null when
deserialized into a java object. This was unforeseen and was left
unchecked when accessing the variable.

Let's fix this by returning an empty string as the default value of
display name in the event of a null case, and also include unit tests
for StandaloneConfig parser.