[#395] Fix unhandled IllegalArgumentException in RepoConfiguration#update (#397)
Standalone config goes through validation check before being utilized
for analysis. In the event where standalone configs contain invalid
value(s), RepoSense will indicate it by throwing an
IllegalArgumentException.

However, in #395, a standalone config with invalid `ignoreCommitList`
was found in one of the repositories which has caused an intermediate
crash during the analysis. This is due to the IllegalArgumentException
not being properly handled.

Let's handle the IllegalArgumentException by ignoring the standalone
config, also, ensure that none of the fields prior to validation.