[#391, #398, #399] Hotfix for outburst of RuntimeException (#400)
Due to the nature of our application, and the lack of simpler
alternatives, our application heavily relies on the external command
line utility, `git`. To make use of the tool, we employed a
CommandRunner class which starts a separate process to execute various
operations.

This, however, encourages the outbreak of RuntimeException as numerous
factors can cause `git` to return a non-zero status code and
CommandRunner allows these exceptions to go unchecked.

To prevent RuntimeException from interrupting the analysis of multiple
repos, let's skip the repo under analysis whenever a RunTimeException
occurs.