[#413] Update Travis-CI configuration to fix pcre2 dependency (#414)
Git in Travis macosx image lacked pcre2 dependency to run checkstyle.
Thus, a line was added to reinstall Git with pcre2.

However, Git is then updated to use pcre2 by default and --with-pcre2
option was removed. This missing option causes Travis to fail.

Let's remove the line to reinstall Git with pcre2 in Travis configuration
to fix the issue.