DeveloperGuide: update documentation for frontend (#242)
The new dashboard doesn't come with the updated development
documentation. And this can prove troublesome for new developers as
they will have to spend a lot of time in the current codebase to make
sense of how the dashboard is structured.

Let's include a short writeup of the dashboard framework to help
explain how the data is being rendered more clearly.