[#393] Remove undefined parameter in the URL (#432)
window.location.hash is initially empty and calling split('&')
creates an array containing a single empty string.

When trying to split this empty string into a key and value,
the result is an empty key and and undefined value which
causes =undefined to appear in the URL.

Let's add a check to ensure that the key is not empty to
prevent this empty string from being included when adding
key value pairs to window.hashParams. This also prevents
any empty keys introduced as a result of user error, perhaps
from editing the URL manually, from being processed.