[#30] Report#Toolbar: rename Team to Repo/Branch (#422)
In our tool bar, we use the term `Team` to define the group of authors
that are from the same repository and branch. 

However, the use of this term can cause confusion to users as it is not
a common terminology among the Git users.

To improve the clarity of our selectable options, let's rename the word
Team to Repo/Branch.