[#199] Clean docs of confusing dates (#415)
The sample dates given in the User Guide and Developer Guide use a day
that is lower than 12.

This makes the day fits the month criteria which can cause confusion to
readers and mixing up the correct date format. This has unnecessary
increase the perplexity of our documentation and needs to be
changed.

Let's change the "dd" part of the dates in the examples to be above 12
so that there will be no bafflement between the day and month fields.