Show breakdown of blank line for total contribution in code panel
With the addition of #385, code view now shows a breakdown of total
contribution by filetype.

However, this does not provide any information on the blank LOC and
non-blank LOC.

Let's show a breakdown of total contribution by blank lines and
non-blank lines in the code view, to provide a better understanding of
the actual amount of code contributed by the author.

Let's also show this breakdown for each file type in the form of a
tooltip.