[#396] Provide a link to repo even when not grouped by repo (#426)
Link to the repository is only available in the chart view when the
`group by repo/branch` is selected, represented by an embedded
hyperlink in the group title.

This causes inconvenience to the users who want to visit the repository
when the authors are listed individually, as the link to repository
cannot be found. 

Let's move the hyperlink in the group title to the individual author's
code view, and also add a button next to every author's name in the
chart view, to ensure consistent navigability to author's repository.