CommitsReporter: add support for per-commit-view (#484)
The dashboard only supports daily and weekly interval views for the
commits.

This is due to commits.json file only holding the daily and weekly
interval contributions summary data.

As part of implementing a per-commit-view on the dahsboard, let's
update commits.json to store the list of each CommitResult inside
the daily interval contribution data. This list of CommitResult with the
full commit hash will also allow for tracking of previously analyzed
commits for incremental build.

As the weekly interval view is now generated on the frontend based
on the daily view by #129, there is also no longer a need to store the
weekly interval contribution data in commits.json.

Let's remove the weekly interval contribution data in commits.json
to reduce the size of the json files.