[#447] CommitInfoAnalyzer: handle empty commit message (#440)
We analyze the raw commit history to obtain information, such as
author, hash, date, message, insertion, deletion etc, about the
commit. The raw data undergoes splitting in order to obtain internal
conceptual representation. However, in the event of encountering a
commit with empty messages, the last index of split elements would be
missing and cause an ArrayIndexOutOfBoundException to be thrown. 

Although, by default, git does not allow empty commit message but this
behavior can be override by this command `git commit
--allow-empty-message`. Also, lightweight IDEs like Visual Studio Code
do allow users to commit without any message by accident.

To prevent breakage due to these corner cases, let's perform an index
check on the elements to avoid RepoSense crash due to the uncaught
exception.

Additional info:
A corner test case on this is included. It is made on the
empty-commit-message branch of testrepo-Alpha.

When committing changes on the Visual Studio Code, if only white spaces
were entered in commit message box, the IDE will commit the changes
without a message.