[#375] Config: add new parameter - author's email (#493)
When merging a pull request on the GitHub platform, GitHub uses the
author's profile name as the Author Name in the merge commit. Unlike
RepoSense which uses Author Name(s) in tracking the authors'
contribution, GitHub uses emails to track the authors' activity. 

This can however pose a problem to our users. When the authors change
their profile name, they will need to update their list of aliases in
the config files correspondingly in order to maintain the trackability
of RepoSense which can be a hassle to our users.

To provide better experience to our users, let's add the email field to
our csv and json configuration files.