[#373] Fix typing lag in the report search bar (#495)
Typing in the report search bar seems to be a bit laggy on the
dashboard-demo. Investigation reveals that lag is caused by the vue
object trying to rewrite the `v_model` on every keystroke.

Let's update and trigger the search only when the typing is complete,
ie we will be updating the summary view as triggered by the `onchange`
event of the search filter `input`.