merging master into per-commit-view before #500 can go in (#507)
* [#350] [Enhancement] author-config.csv: make repo's location optional field  (#441)

In the author-config.csv, each author must provide a location field,
which is mapped to the repository to be analyzed in repo-config.csv.

However, the user may want to analyze the contribution of an author for
all the repositories in repo-config.csv, and it is inconvenient for the
user to type it out for each repository.

Let's allow users to leave the location field empty in
author-config.csv, which will add this author to all the repositories
in repo-config.csv.

Things to note, if an author with empty location exist, all the repos
will not use the default all authors as this author with empty location
will beadded to all the repositories.

* [#474] Format: add 'txt' format to list of default formats (#475)

Format#DEFAULT_FORMAT_STRINGS contains a list of common file extensions
within git repositories, which RepoSense will use to run the analysis by default
if no formats is given by the user.

However, this list does not contain the `txt` file extension, which is
also another commonly used file extension in git repositories.

For user's convenience, let's add the `txt` file extension to the list
of default format strings as well.

* [#447] CommitInfoAnalyzer: handle empty commit message (#440)

We analyze the raw commit history to obtain information, such as
author, hash, date, message, insertion, deletion etc, about the
commit. The raw data undergoes splitting in order to obtain internal
conceptual representation. However, in the event of encountering a
commit with empty messages, the last index of split elements would be
missing and cause an ArrayIndexOutOfBoundException to be thrown. 

Although, by default, git does not allow empty commit message but this
behavior can be override by this command `git commit
--allow-empty-message`. Also, lightweight IDEs like Visual Studio Code
do allow users to commit without any message by accident.

To prevent breakage due to these corner cases, let's perform an index
check on the elements to avoid RepoSense crash due to the uncaught
exception.

Additional info:
A corner test case on this is included. It is made on the
empty-commit-message branch of testrepo-Alpha.

When committing changes on the Visual Studio Code, if only white spaces
were entered in commit message box, the IDE will commit the changes
without a message.

* [#481] Remove unused report folder (#489)

The report folder in the project root seems to be unused.

Let's remove the unused folder keep the project clean.

* [#375] Config: add new parameter - author's email (#493)

When merging a pull request on the GitHub platform, GitHub uses the
author's profile name as the Author Name in the merge commit. Unlike
RepoSense which uses Author Name(s) in tracking the authors'
contribution, GitHub uses emails to track the authors' activity. 

This can however pose a problem to our users. When the authors change
their profile name, they will need to update their list of aliases in
the config files correspondingly in order to maintain the trackability
of RepoSense which can be a hassle to our users.

To provide better experience to our users, let's add the email field to
our csv and json configuration files.

* [#373] Fix typing lag in the report search bar (#495)

Typing in the report search bar seems to be a bit laggy on the
dashboard-demo. Investigation reveals that lag is caused by the vue
object trying to rewrite the `v_model` on every keystroke.

Let's update and trigger the search only when the typing is complete,
ie we will be updating the summary view as triggered by the `onchange`
event of the search filter `input`.

* [#297] UserGuide: add FAQ section to address common problems (#494)

Our UserGuide does not have a FAQ section.

Let's add a FAQ section in our UserGuide to address common questions
and issues, allowing users troubleshoot them themselves.

* [#399] Fix RuntimeException due to ambiguous argument in git rev-list (#505)

When a file with the same name as the analyzing branch exist in the
repo, an uncaught RuntimeException will occur due to ambiguous argument
in git rev-list causing RepoSense to crash.

Let's add the `--` revision-path separator in `git rev-list` command to
distinguish the branch name from the ambiguous argument.

* updated test

* undo travis