[#392] Show a 'loading...' message while loading the report for the first time (#498)
RepoSense needs to load the resource files before it can generate
the dashboard, which can take a significant amount of time if the
resource files given are large.

However, RepoSense remains on the initial page, with the option to
upload a zip file, even after it has started loading these resource
files. This can be misleading to users as they may think they have
selected the wrong file and need to upload it again.

There is also no indication given if an invalid zip file was given,
which may make users wait indefinitely for the dashboard to load when
it won't.

Let's implement a loading page to inform the user that the application
is loading and a pop-up to inform the user if an invalid zip file is
given.