Make center divider adjustable #403 (#418)
RepoSense has a code panel and summary panel, each occupies
half of the screen. However, the code panel width cannot be resized.

Enabling resizing of the code panel will allow user to adjust it to
their comfort.

Let's have an adjustable code panel by dragging the width of the panel.