[#502] Enable ESLint JavaScript linter on travis (#514)
ESLint was implemented, however it doesn't seems to be enforcing
style check on any of the CI platforms. 

This is due to the glob that was passed into eslint was incorrect which
causes it to skip the js files.

Let's modify the glob string to enable the javascript linter, and also
fixed the violations that were found.