[#527] Improve highlightjs usage to prevent report from lagging/crashing (#533)
hightlightjs (`hljs`) is used for highlighting of codes in code view
giving it context and readability. It can be triggered to render an
element with `window.hljs.highlightBlock(element)`.

`hljs` operations are expensive and stateless, where it re-renders even
on a rendered DOM and uses automatic detection of programming languages
if the language is not specified. Also, multiple `hljs` operations can
be triggered at once, causing the report to lag and crash.
Additionally, it seems to render regardless of trigger for any code in
`<pre><code>`. Furthermore, in our frontend, `vuejs` will trigger `hljs`
in `Vue#update`, when there is a data change, causing DOM to be 
re-rendered.

Let's optimize hljs usage, by triggering it when code is inserted into
the DOM, enclosing the code with a different tag, specifiying the
language to prevent expensive auto detection.