[#468] Add per-commit granularity to the report (#500)
The dashboard only allows summary view for daily and weekly
granularities.

While the daily and weekly intervals give the reviewer the rough
impression of the authors' workload, it does not provide any insights
on the authors' commit frequency, commit organization and merged
changes. These information are important in helping users to further
understand the authors' workflow and contribution.

Let's implement summary view for per commit granularity. To do so,
let's update `commits.json` report to store the list of each
CommitResult inside the daily interval contribution data, and update
the frontend implement the per commit granularity using this data.

As the weekly interval view is now generated on the frontend based
on the daily view by #129, there is no longer a need to store the
weekly interval contribution data in commits.json.

Let's also remove the weekly interval contribution data in
`commits.json` report to reduce the size of the json files.