Remove redundant commands in Travis config (#551)
In the Travis config file, we specify our Linux environment to use
JDK9 with `jdk: oraclejdk9` and `addons.apt.packages key`, but the
latter is redundant. Similarly, the command used to install frontend
npm dependencies can be simpified. The cd command does not change the
install location of the npm dependencies as frontend folder does not
contain a package.json file.

Let's remove the redundant commands for clarity and maintainability.