[#521] CodeView: Collapse/Expand All button not updated correctly after hiding all files (#529)
If the select-all box for file formats in the code panel is unchecked,
the 'Collapse/Expand All' button will change to become 'Expand All' as
all files are hidden. However, when the select-all box is checked
again, all the files rendered will be expanded, but the
'Collapse/Expand All' button will remain at 'Expand All'.

This is due to files being automatically expanded when they are
rendered, after we check any of the file format checkboxes.

Let's update the 'Collapse/Expand All' button to also track files
that are newly rendered that are expanded whenever a file format 
checkbox is clicked.