[#449, #479] ArgsParser: extract magic strings and follow a widely-used standard for CLI (#526)
In ArgsParser, the string flags used for the CLI are magic strings
repeated in multiple places. Meanwhile, the standard of CLI argument
used does not follow a widely-used standard.

We should avoid the use of magic strings, instead defining them in a
central location. As RepoSense is designed for CLI usage, it is also
a good idea to follow a widely-used standard for CLI to match user
expectations. 

Let's define the magic strings in a central location as constants and
follow the GNU Coding Standard[1] for the CLI, using `--` prefix for
`long-named flags` and `-` for its equivalent `single-letter alias`.

[1] GNU Coding Standards: Command Line Interfaces
https://www.gnu.org/prep/standards/html_node/Command_002dLine-Interfaces.html