[#557] Travis: check scripts do not work in Travis macOS environment (#559)
Following #414, the line to reinstall `git` with `pcre2` was removed in
Travis macOS environment. As running `brew reinstall --with-pcre2 git`
would caused Travis to fail. Furthermore, `pcre2` now ships with
`git >= 2.19`.

Since its removal, check scripts, that depends on `pcre2`, do not work
but Travis macOS still passes. This is due to the use of an older
version of git which does not comes with `pcre2`.

Let's use the latest version of Travis macOS image, that comes with git
2.19.1, to ensure check scripts run correctly.