[#546] FAQ: add another entry to propose email as alternative to GitHub ID (#568)
In the UserGuide FAQ section, we mention that commit contribution
may not appear in the ramp chart if the user did not associate the
correct author name aliases to their GitHub ID, and to double check
their author name aliases.

After #493 though, RepoSense allows user to associate their GitHub
email to their GitHub ID as well, which is more accurate as it is a
fixed value, compared to author name aliases which can have many
different values depending on their local git settings and on different
machines. As such, it is better to ask users to use email instead of
author name aliases if they face such problems of their contribution
not being tracked by RepoSense.

Let's update FAQ to propose email as an alternative to associate to the
user's GitHub ID instead.

The command given in the `A Note About Git Author Name` is not
exactly accurate, as it only changes the git author name settings for
future initiated or cloned git repositories, but not for the current
git repository, which is what most users who are reading the FAQ would
want.

Let's also update the `A Note About Git Author Name` section to
include the correct command for changing the git author name settings
for the current git repository as well.