[#585] Fix linearly unsynchronous ramp charts (#591)
Checks are performed on the dates prior to every update of the ramp
chart's filter.

However, the checks only accounts for cases where the dates are valid,
checking that they are within the given range. When an invalid date
such as "2019-02-31" is given, the checks are bypassed, causing
ramp charts of different repositories to have different end dates and
linearly unsynchronous.

Let's add an additional check to ensure that valid dates were
given before we update the filter fields.