[#456] ChartView: show breakdown of file formats in Contribution Bar (#487)
Contribution Bar is a single red bar which represents total
contribution of the author, without any additional information.

Providing a breakdown by file format in the Contribution Bar allows
users to view the file format used in a repository, and compare
contributions by file formats with a quick glance at the Chart Panel,
without having to open the Code Panel.

Let's reflect the breakdown of file formats in Contribution Bar.

The user can hover over a specific section of the Contribution Bar
to see information about the file formats, total lines contributed
and analysis period in the tooltip.

The user can refer to the legend below a repo name in the chart view
to determine which color represents which file formats.

If the user has specified file formats, the Contribution Bar and legend
reflects only those file formats, in order specified by user.
Otherwise, the file formats are sorted alphabetically.

Since this file format information is being processed in the front-end
in code view (v_authorship), let's move the processing of file format
information to the backend in order to be able to use the pre-processed
file format information for both chart view (v_summary) and code view
(v_authorship). Let's also refactor the file format processing in
code-view (v_authorship) to match the changes being made.