Refactor toolbar into v-authorship tab (#566)
The toolbar for the "expand/collapse all" text is outside of the
v-authorship tab.

This shouldn't be the case as it makes it difficult to extend and
include other kinds of tabs.

Let's shift the toolbar into the v-authorship tab, and make some
corresponding changes to how the toggle state is changed.

This will keep the toolbar within the authorship tab, which makes sense
as the button is only relevant for the authorship tab. In light that we
are going to have other kinds of tabs to extend the use of the side tab
in the future.